Rust::com update the Runtime and FFI Implementation and remove the old APIs
* Removed the Rust Old APIs implementation and target
* Updated the re-used code into FFI folder
