+++ score/mw/com/impl/rust/com-api/com-api-ffi-lola/BUILD
@@ -28,7 +28,7 @@ cc_library(
         "//score/mw/com:runtime_configuration",
         "//score/mw/com:types",
         "//score/mw/com/impl/plumbing:sample_ptr",
-        "//score/mw/com/impl/rust:proxy_bridge_cpp",
+        "@score_baselibs//score/language/futurecpp",
     ],
 )
 
@@ -42,7 +42,6 @@ rust_library(
     deps = [
         ":registry_bridge_macro_cpp",
         "//score/mw/com/impl/plumbing/rust:sample_ptr_rs",
-        "//score/mw/com/impl/rust:mw_com",
     ],
 )
