+++ score/mw/com/impl/rust/com-api/com-api-ffi-lola/registry_bridge_macro.h
@@ -68,20 +68,119 @@
 #include "score/mw/com/impl/proxy_base.h"
 #include "score/mw/com/impl/proxy_event.h"
 #include "score/mw/com/impl/proxy_event_base.h"
-#include "score/mw/com/impl/rust/proxy_bridge.h"
 #include "score/mw/com/impl/skeleton_base.h"
 #include "score/mw/com/impl/skeleton_event.h"
 #include "score/mw/com/impl/skeleton_event_base.h"
 #include "score/mw/com/types.h"
 
+#include <score/assert.hpp>
+#include <utility>
 #include <cstdint>
 #include <memory>
 #include <string_view>
 #include <unordered_map>
-#include <utility>
 
 namespace score::mw::com::impl::rust
 {
+
+struct FatPtr
+{
+    const void* vtbl;
+    void* data;
+};
+
+template <typename R, typename... Args>
+class RustBoxedCallable;
+
+template <typename R, typename... Args>
+class RustRefMutCallable;
+
+template <template <typename, typename...> class FnMutHandler, typename R, typename... Args>
+class RustFnMutCallableBase
+{
+  public:
+    using Handler = FnMutHandler<R, Args...>;
+
+    explicit RustFnMutCallableBase(const FatPtr& dyn_fnmut) noexcept : ptr_{dyn_fnmut}
+    {
+        SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD_MESSAGE(dyn_fnmut.data != nullptr,
+                                                    "Failed creating a RustFnMutCallable due to an invalid pointer");
+    }
+
+    RustFnMutCallableBase(const RustFnMutCallableBase&) = delete;
+    RustFnMutCallableBase(RustFnMutCallableBase&& other) noexcept : ptr_{other.ptr_}
+    {
+        other.ptr_.data = nullptr;
+    }
+
+    RustFnMutCallableBase& operator=(const RustFnMutCallableBase&) = delete;
+    RustFnMutCallableBase& operator=(RustFnMutCallableBase&& other) noexcept
+    {
+        std::swap(ptr_, other.ptr_);
+        return *this;
+    }
+
+    ~RustFnMutCallableBase() noexcept
+    {
+        if (ptr_.data != nullptr)
+        {
+            Handler::dispose(ptr_);
+        }
+    }
+
+  protected:
+    void CheckValid() const noexcept
+    {
+        SCORE_LANGUAGE_FUTURECPP_ASSERT_PRD_MESSAGE(ptr_.data != nullptr,
+                                                    "Tried to use a Rust FnMut callable without a valid pointer");
+    }
+
+    FatPtr ptr_;
+};
+
+template <template <typename, typename...> class FnMutHandler, typename R = void, typename... Args>
+class RustFnMutCallable : public RustFnMutCallableBase<FnMutHandler, R, Args...>
+{
+  public:
+    using Base = RustFnMutCallableBase<FnMutHandler, R, Args...>;
+    using RustFnMutCallableBase<FnMutHandler, R, Args...>::RustFnMutCallableBase;
+
+    template <typename... CallArgs>
+    R operator()(CallArgs&&... call_args) noexcept
+    {
+        Base::CheckValid();
+        return Base::Handler::invoke(this->ptr_, std::forward<CallArgs>(call_args)...);
+    }
+
+    template <typename... CallArgs>
+    R operator()(CallArgs&&... call_args) const noexcept
+    {
+        Base::CheckValid();
+        return Base::Handler::invoke(this->ptr_, std::forward<CallArgs>(call_args)...);
+    }
+};
+
+template <template <typename, typename...> class FnMutHandler, typename... Args>
+class RustFnMutCallable<FnMutHandler, void, Args...> final : public RustFnMutCallableBase<FnMutHandler, void, Args...>
+{
+  public:
+    using Base = RustFnMutCallableBase<FnMutHandler, void, Args...>;
+    using RustFnMutCallableBase<FnMutHandler, void, Args...>::RustFnMutCallableBase;
+
+    template <typename... CallArgs>
+    void operator()(CallArgs&&... call_args) noexcept
+    {
+        Base::CheckValid();
+        Base::Handler::invoke(this->ptr_, std::forward<CallArgs>(call_args)...);
+    }
+
+    template <typename... CallArgs>
+    void operator()(CallArgs&&... call_args) const noexcept
+    {
+        Base::CheckValid();
+        Base::Handler::invoke(this->ptr_, std::forward<CallArgs>(call_args)...);
+    }
+};
 
 /// \brief String view for FFI boundary passing strings from Rust to C++
 /// \details Holds a pointer to a string and its length without requiring null termination.
