+++ score/mw/com/impl/rust/com-api/com-api-runtime-lola/BUILD
@@ -28,7 +28,6 @@ rust_library(
     deps = [
         "//score/mw/com/impl/plumbing/rust:sample_allocatee_ptr_rs",
         "//score/mw/com/impl/plumbing/rust:sample_ptr_rs",
-        "//score/mw/com/impl/rust:mw_com",
         "//score/mw/com/impl/rust/com-api/com-api-concept",
         "//score/mw/com/impl/rust/com-api/com-api-ffi-lola:bridge_ffi_lola",
         "//score/mw/com/impl/rust/com-api/com-api-ffi-lola:bridge_ffi_rs",
