--- score/mw/com/impl/rust/proxy_bridge_macros.h
@@ -1,16 +1,0 @@
-/********************************************************************************
- * Copyright (c) 2025 Contributors to the Eclipse Foundation
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Apache License Version 2.0 which is available at
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * SPDX-License-Identifier: Apache-2.0
- ********************************************************************************/
-
-// This header is provided for backwards compatibility and can be removed once everyone migrated to bridge_macros.h
-
-#include "score/mw/com/impl/rust/bridge_macros.h"
