--- score/mw/com/impl/rust/skeleton_bridge.rs
@@ -1,103 +1,0 @@
-/********************************************************************************
- * Copyright (c) 2025 Contributors to the Eclipse Foundation
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Apache License Version 2.0 which is available at
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * SPDX-License-Identifier: Apache-2.0
- ********************************************************************************/
-use std::sync::Arc;
-
-mod ffi {
-    use std::marker::PhantomData;
-
-    /// This type represents `score::mw::com::impl::SkeletonWrapperClass` as an opaque struct for any
-    /// template argument, as the type isn't relevant when dealing with it as an opaque type.
-    /// Note that this struct is empty as we only use references to it on Rust side.
-    #[repr(C)]
-    pub struct SkeletonWrapperClass {
-        _dummy: [u8; 0],
-    }
-
-    /// This type represents `score::mw::com::impl::SkeletonEventBase` as an opaque struct.
-    /// Note that this struct is empty as we only use references to it on Rust side.
-    #[repr(C)]
-    pub struct SkeletonEvent<T> {
-        _dummy: [u8; 0],
-        _data: PhantomData<T>,
-    }
-}
-
-pub use ffi::SkeletonEvent as NativeSkeletonEvent;
-pub use ffi::SkeletonWrapperClass;
-
-pub enum UnOffered {}
-pub enum Offered {}
-pub trait OfferState {}
-impl OfferState for UnOffered {}
-impl OfferState for Offered {}
-
-pub trait SkeletonOps: Sized {
-    /// Sends the skeleton event.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the event cannot be sent.
-    #[allow(clippy::result_unit_err)]
-    // Suppressing clippy::result_unit_err as the common error type is used in upper layer when error occurs.
-    fn send(&self, event: *mut ffi::SkeletonEvent<Self>) -> common::Result<()>;
-}
-
-pub struct SkeletonEvent<T: SkeletonOps, S: OfferState, L> {
-    event: *mut ffi::SkeletonEvent<T>,
-    skeleton: Arc<L>,
-    _marker: std::marker::PhantomData<S>,
-}
-
-impl<T: SkeletonOps, S: OfferState, L> SkeletonEvent<T, S, L> {
-    /// Creates a new `SkeletonEvent`.
-    pub fn new(event: *mut NativeSkeletonEvent<T>, skeleton: Arc<L>) -> Self {
-        Self {
-            event,
-            skeleton,
-            _marker: std::marker::PhantomData,
-        }
-    }
-}
-
-impl<T: SkeletonOps, L> SkeletonEvent<T, UnOffered, L> {
-    /// Offers the skeleton event.
-    pub fn offer(self) -> SkeletonEvent<T, Offered, L> {
-        SkeletonEvent::<T, Offered, L> {
-            event: self.event,
-            skeleton: self.skeleton,
-            _marker: std::marker::PhantomData,
-        }
-    }
-}
-
-impl<T: SkeletonOps, L> SkeletonEvent<T, Offered, L> {
-    #[allow(clippy::result_unit_err)]
-    // Suppressing clippy::result_unit_err as the common error type is used in upper layer when error occurs.
-    /// Sends the skeleton event.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the event cannot be sent.
-    pub fn send(&self, stamped_data: T) -> common::Result<()> {
-        stamped_data.send(self.event)
-    }
-
-    /// Stops offering the skeleton event.
-    pub fn stop_offer(self) -> SkeletonEvent<T, UnOffered, L> {
-        SkeletonEvent::<T, UnOffered, L> {
-            event: self.event,
-            skeleton: self.skeleton,
-            _marker: std::marker::PhantomData,
-        }
-    }
-}
